@@ -1,16 +1,16 @@
 declare module '*.svg' {
-    import React = require('react');
-    export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>;
-    const src: string;
-    export default src;
+	import React = require('react');
+	export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>;
+	const src: string;
+	export default src;
 }
 
 declare module '*.jpg' {
-    const content: string;
-    export default content;
+	const content: string;
+	export default content;
 }
 
 declare module '*.png' {
-    const content: string;
-    export default content;
+	const content: string;
+	export default content;
 }
