@@ -1,1 +1,1 @@
-declare module '*.module.css';
+declare module '*.module.scss';
