@@ -1,51 +1,54 @@
 import { ReactComponent as IconReply } from 'assets/icon-reply.svg';
 import { ReactComponent as IconDelete } from 'assets/icon-delete.svg';
 import { ReactComponent as IconEdit } from 'assets/icon-edit.svg';
+import { Button } from 'components/UI';
+import styles from './CommentActions.module.scss';
 
-import Button from 'components/UI/Button';
+interface Props {
+	isCurrentUser: boolean;
+	onCommentDelete: () => void;
+	onCommentEdit: () => void;
+	onCommentReply: () => void;
+}
 
-import styles from './CommentActions.module.css';
+const CommentActions: React.FC<Props> = ({
+	isCurrentUser,
+	onCommentDelete,
+	onCommentEdit,
+	onCommentReply,
+}) => {
+	const authorizedActionsContent = (
+		<>
+			<Button
+				variant="outline-danger"
+				iconElement={<IconDelete />}
+				label="Delete"
+				onClick={onCommentDelete}
+			/>
 
-type Props = {
-    isCurrentUser: boolean;
-    onCommentReply: () => void;
-    onCommentEdit: () => void;
-    onCommentDelete: () => void;
-};
+			<Button
+				variant="outline-primary"
+				iconElement={<IconEdit />}
+				label="Edit"
+				onClick={onCommentEdit}
+			/>
+		</>
+	);
 
-const CommentActions: React.FC<Props> = props => {
-    const globalUserActions = (
-        <Button
-            variant="outline-primary"
-            label="Reply"
-            iconElement={<IconReply />}
-            onClick={props.onCommentReply}
-        />
-    );
+	const globalActionsContent = (
+		<Button
+			variant="outline-primary"
+			iconElement={<IconReply />}
+			label="Reply"
+			onClick={onCommentReply}
+		/>
+	);
 
-    const currentUserActions = (
-        <>
-            <Button
-                variant="outline-danger"
-                label="Delete"
-                iconElement={<IconDelete />}
-                onClick={props.onCommentDelete}
-            />
-
-            <Button
-                variant="outline-primary"
-                label="Edit"
-                iconElement={<IconEdit />}
-                onClick={props.onCommentEdit}
-            />
-        </>
-    );
-
-    return (
-        <div className={styles.commentActions}>
-            {props.isCurrentUser ? currentUserActions : globalUserActions}
-        </div>
-    );
+	return (
+		<div className={styles.actions}>
+			{isCurrentUser ? authorizedActionsContent : globalActionsContent}
+		</div>
+	);
 };
 
 export default CommentActions;
