@@ -1,31 +1,29 @@
-import Modal from 'components/UI/Modal';
-import Button from 'components/UI/Button';
+import { Modal, Button } from 'components/UI';
+import styles from './CommentDeleteModal.module.scss';
 
-import styles from './CommentDeleteModal.module.css';
+interface Props {
+	onConfirm: () => void;
+	onCancel: () => void;
+}
 
-type Props = {
-    onCancel: () => void;
-    onConfirm: () => void;
-};
+const CommentDeleteModal: React.FC<Props> = ({ onConfirm, onCancel }) => {
+	return (
+		<Modal onClose={onCancel}>
+			<div className={styles.deleteModalContent}>
+				<h2>Delete Comment</h2>
 
-const CommentDeleteModal: React.FC<Props> = props => {
-    return (
-        <Modal onClose={props.onCancel}>
-            <div className={styles.deleteModalContent}>
-                <h2>Delete comment</h2>
+				<p>
+					Are you sure you want to delete this comment? This will remove the comment and
+					can't be undone.
+				</p>
 
-                <p>
-                    Are you sure you want to delete this comment? This will remove the comment and can't be
-                    undone.
-                </p>
-
-                <div className={styles.deleteModalActions}>
-                    <Button variant="contained-secondary" label="No, Cancel" onClick={props.onCancel} />
-                    <Button variant="contained-danger" label="Yes, Delete" onClick={props.onConfirm} />
-                </div>
-            </div>
-        </Modal>
-    );
+				<div className={styles.deleteModalActions}>
+					<Button variant="contained-secondary" label="No, Cancel" onClick={onCancel} />
+					<Button variant="contained-danger" label="Yes, Delete" onClick={onConfirm} />
+				</div>
+			</div>
+		</Modal>
+	);
 };
 
 export default CommentDeleteModal;
