@@ -1,26 +1,28 @@
-import styles from './CommentHeader.module.css';
+import { IUser } from 'models';
+import { CommentTime } from 'components/Comments';
+import styles from './CommentHeader.module.scss';
 
-type Props = {
-    username: string;
-    createdAt: string;
-    isCurrentUser: boolean;
-};
+interface Props {
+	user: IUser;
+	createdAt: string;
+	isCurrentUser: boolean;
+}
 
-const CommentHeader: React.FC<Props> = props => {
-    const userImg = require(`assets/avatars/image-${props.username}.png`);
+const CommentHeader: React.FC<Props> = ({ user, createdAt, isCurrentUser }) => {
+	const userImg = require(`assets/avatars/${user.image.png}`);
 
-    return (
-        <div className={styles.commentHeader}>
-            <img src={userImg} alt={props.username} className={styles.userImg} />
+	return (
+		<header className={styles.header}>
+			<img src={userImg} alt={user.username} />
 
-            <div className={styles.usernameWrapper}>
-                <p className={styles.username}>{props.username}</p>
-                {props.isCurrentUser && <span className={styles.badge}>you</span>}
-            </div>
+			<div className={styles.usernameWrapper}>
+				<p className={styles.username}>{user.username}</p>
+				{isCurrentUser && <span className={styles.badge}>you</span>}
+			</div>
 
-            <p>{props.createdAt}</p>
-        </div>
-    );
+			<CommentTime date={createdAt} />
+		</header>
+	);
 };
 
 export default CommentHeader;
