@@ -1,60 +1,41 @@
-import { useRef, useEffect } from 'react';
-
+import { useRef } from 'react';
 import { motion } from 'framer-motion';
+import { Input, Button } from 'components/UI';
+import { IUser } from 'models';
+import styles from './CommentReplyForm.module.scss';
 
-import Input from 'components/UI/Input';
-import Button from 'components/UI/Button';
+interface Props {
+	currentUser: IUser;
+	onFormSubmit: (value: string) => void;
+}
 
-import { useComments } from 'context';
-import { deepFindCommentReplies } from 'utils';
-import { CommentReply } from 'models';
+const CommentReplyForm: React.FC<Props> = ({ currentUser, onFormSubmit }) => {
+	const textInputRef = useRef<HTMLTextAreaElement>(null);
 
-import styles from './CommentReplyForm.module.css';
+	const userImg = require(`assets/avatars/image-${currentUser.username}.png`);
 
-const CommentReplyForm = () => {
-    const replyInputRef = useRef<HTMLTextAreaElement>(null);
-    const commentsCtx = useComments();
+	const handleFormSubmit = (event: React.FormEvent) => {
+		event.preventDefault();
 
-    const userImg = require(`assets/avatars/image-${commentsCtx.currentUser}.png`);
+		const textValue = textInputRef.current!.value;
+		if (!textValue.trim().length) return;
 
-    const handleFormSubmit = (event: React.FormEvent) => {
-        event.preventDefault();
+		onFormSubmit(textValue);
+	};
 
-        const replyContent = replyInputRef.current!.value;
-        if (!replyContent.trim().length) return;
-
-        const replyingToComment = deepFindCommentReplies(
-            commentsCtx.commentReplyId as number,
-            commentsCtx.comments
-        );
-
-        const commentReply = new CommentReply(
-            commentsCtx.currentUser,
-            replyingToComment.user.username,
-            replyContent
-        );
-
-        commentsCtx.replyComment(commentReply);
-        commentsCtx.setCommentReplyId(null);
-    };
-
-    useEffect(() => {
-        replyInputRef.current!.focus();
-    }, []);
-
-    return (
-        <motion.form
-            className={styles.replyForm}
-            onSubmit={handleFormSubmit}
-            initial={{ y: -50, opacity: 0.6 }}
-            animate={{ y: 0, opacity: 1 }}
-            transition={{ duration: 0.5 }}
-        >
-            <img src={userImg} alt={commentsCtx.currentUser} />
-            <Input ref={replyInputRef} input={{ placeholder: 'Add a reply...' }} />
-            <Button variant="contained-primary" label="Reply" />
-        </motion.form>
-    );
+	return (
+		<motion.form
+			className={styles.replyForm}
+			onSubmit={handleFormSubmit}
+			initial={{ y: -50, opacity: 0.6, scale: 0.6 }}
+			animate={{ y: 0, opacity: 1, scale: 1 }}
+			transition={{ duration: 0.3 }}
+		>
+			<img src={userImg} alt={currentUser.username} />
+			<Input ref={textInputRef} input={{ placeholder: 'Add a reply...' }} />
+			<Button variant="contained-primary" label="Reply" />
+		</motion.form>
+	);
 };
 
 export default CommentReplyForm;
