@@ -1,12 +1,12 @@
 import { forwardRef } from 'react';
-import styles from './Input.module.css';
+import styles from './Input.module.scss';
 
-type Props = {
-    input?: React.TextareaHTMLAttributes<HTMLTextAreaElement>;
-};
+interface Props {
+	input?: React.TextareaHTMLAttributes<HTMLTextAreaElement>;
+}
 
 const Input = forwardRef<HTMLTextAreaElement, Props>((props, ref) => (
-    <textarea ref={ref} className={styles.textarea} rows={3} {...props.input} />
+	<textarea ref={ref} className={styles.textarea} rows={3} {...props.input} />
 ));
 
 export default Input;
